# specs/001-map-world-points/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Points on the Map API
  version: 0.1.0
  description: >
    Express REST API for 001-map-world-points. All authenticated routes
    require Clerk-issued session or Bearer token (per deployment).
servers:
  - url: /api
    description: API base (adjust per gateway)

tags:
  - name: public
  - name: points
  - name: folders
  - name: tags
  - name: me
  - name: meta
  - name: auth
    description: Clerk session probes and inbound webhook for user sync (US2).

components:
  securitySchemes:
    clerkBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk session JWT

  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: VALIDATION_ERROR }
        message: { type: string }
        details: { type: object, additionalProperties: true }

    Point:
      type: object
      required: [id, title, latitude, longitude, createdAt, visibility]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        description: { type: string, nullable: true }
        photoUrl: { type: string, format: uri, nullable: true }
        latitude: { type: number, format: double, minimum: -90, maximum: 90 }
        longitude: { type: number, format: double, minimum: -180, maximum: 180 }
        createdAt: { type: string, format: date-time }
        authorId: { type: string, format: uuid }
        visibility: { type: string, enum: [public, group_only] }
        groupId: { type: string, format: uuid, nullable: true }
        folderId: { type: string, format: uuid, nullable: true }
        averageRating: { type: number, nullable: true }
        myRating: { type: integer, minimum: 1, maximum: 5, nullable: true }

    PointCreate:
      type: object
      required: [title, latitude, longitude]
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 5000, nullable: true }
        latitude: { type: number, format: double, minimum: -90, maximum: 90 }
        longitude: { type: number, format: double, minimum: -180, maximum: 180 }
        folderId: { type: string, format: uuid, nullable: true }
        visibility: { type: string, enum: [public, group_only], default: public }
        groupId: { type: string, format: uuid, nullable: true }
        tagIds: { type: array, items: { type: string, format: uuid } }

    Comment:
      type: object
      properties:
        id: { type: string, format: uuid }
        body: { type: string, maxLength: 4000 }
        authorId: { type: string, format: uuid }
        displayName: { type: string }
        createdAt: { type: string, format: date-time }

    PhotoUpload:
      type: object
      description: Signed direct-upload parameters for Cloudinary (multipart POST).
      required:
        [uploadUrl, apiKey, timestamp, signature, uploadPreset, cloudName, folder, photoKey]
      properties:
        uploadUrl:
          type: string
          format: uri
          description: POST URL for the image upload (`https://api.cloudinary.com/v1_1/<cloud>/image/upload`).
        apiKey: { type: string, description: Cloudinary API key (safe to send to an authenticated client for upload only). }
        timestamp: { type: integer, description: Unix seconds included in the signature. }
        signature: { type: string, description: Server-side HMAC for the signed upload_preset and folder. }
        uploadPreset: { type: string }
        cloudName: { type: string }
        folder: { type: string, description: Target folder / public_id prefix (must match signed parameters). }
        photoKey:
          type: string
          description: Cloudinary `public_id` to persist on `points.photo_key` after the client upload completes successfully.

    PointUpdate:
      type: object
      minProperties: 1
      description: Partial update; at least one field required.
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 5000, nullable: true }
        folderId: { type: string, format: uuid, nullable: true }
        visibility: { type: string, enum: [public, group_only] }
        groupId: { type: string, format: uuid, nullable: true }
        tagIds: { type: array, items: { type: string, format: uuid } }
        photoKey:
          type: string
          minLength: 1
          maxLength: 2048
          nullable: true
          description: Must match the canonical public_id from photo-upload (single image per point).

    Folder:
      type: object
      required: [id, name, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        groupId: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }

    FolderCreate:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        groupId: { type: string, format: uuid, nullable: true }

    FolderUpdate:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }

    Tag:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }

    TagCreate:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }

paths:
  /health:
    get:
      tags: [meta]
      summary: Liveness
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }

  /me:
    get:
      tags: [auth]
      security: [{ clerkBearer: [] }]
      summary: >
        Sanity probe — confirms the bearer token resolves to a Clerk session.
        Full user/preferences API is added in T049 (Phase 6).
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                type: object
                required: [authenticated]
                properties:
                  authenticated: { type: boolean, example: true }
                  clerkUserId: { type: string, nullable: true, example: user_2abc... }
        '401':
          description: Missing or invalid Clerk session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /webhooks/clerk:
    post:
      tags: [auth]
      summary: Clerk → backend user-sync webhook (Svix-signed)
      description: >
        Receives `user.created` / `user.updated` / `user.deleted` events from
        Clerk and upserts the internal `users` row keyed by `clerk_id`.
        Authenticated by Svix headers (`svix-id`, `svix-timestamp`,
        `svix-signature`); the `CLERK_WEBHOOK_SECRET` env var holds the
        signing secret. Other event types are accepted with `ignored: true`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Raw Clerk webhook event (verified server-side).
              additionalProperties: true
      responses:
        '200':
          description: Event accepted (or ignored for unhandled types).
          content:
            application/json:
              schema:
                type: object
                required: [received]
                properties:
                  received: { type: boolean, example: true }
                  type: { type: string, example: user.created }
                  ignored: { type: boolean, example: false }
                  clerkId: { type: string, example: user_2abc... }
        '400':
          description: Invalid Svix signature or payload (no DB writes performed).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          description: Webhook secret not configured on the server.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /public/latest:
    get:
      tags: [public]
      summary: Latest five public points (guest + signed-in list widget)
      responses:
        '200':
          description: Up to 5 public points, newest first
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Point' } }

  /public/points/{pointId}:
    get:
      tags: [public]
      summary: Public point read-only detail (guest or signed-in)
      parameters:
        - name: pointId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Point'
        '404': { description: Not found or not visible }

  /map/public:
    get:
      tags: [public]
      summary: Public points in bounding box (signed-in; guest uses latest subset client-side)
      security: [{ clerkBearer: [] }]
      parameters:
        - in: query
          name: southWestLat
          schema: { type: number }
        - in: query
          name: southWestLng
          schema: { type: number }
        - in: query
          name: northEastLat
          schema: { type: number }
        - in: query
          name: northEastLng
          schema: { type: number }
        - in: query
          name: limit
          schema: { type: integer, default: 500, maximum: 2000 }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Point' } }

  /me/preferences:
    get:
      security: [{ clerkBearer: [] }]
      tags: [me]
      summary: Active private group, etc.
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  activeGroupId: { type: string, format: uuid, nullable: true }
    patch:
      security: [{ clerkBearer: [] }]
      tags: [me]
      summary: Set active private group
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                activeGroupId: { type: string, format: uuid, nullable: true }
      responses:
        '200': { description: Updated }
        '403': { description: Not a member of group }

  /folders:
    get:
      security: [{ clerkBearer: [] }]
      tags: [folders]
      summary: List organization folders for the current user
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Folder' } }
    post:
      security: [{ clerkBearer: [] }]
      tags: [folders]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FolderCreate' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Folder' }

  /folders/{folderId}:
    patch:
      security: [{ clerkBearer: [] }]
      tags: [folders]
      parameters:
        - name: folderId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FolderUpdate' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Folder' }
        '404': { description: Not found }
    delete:
      security: [{ clerkBearer: [] }]
      tags: [folders]
      parameters:
        - name: folderId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Deleted }
        '404': { description: Not found }

  /tags:
    get:
      security: [{ clerkBearer: [] }]
      tags: [tags]
      summary: List global tags and current user's tags
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Tag' } }
    post:
      security: [{ clerkBearer: [] }]
      tags: [tags]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TagCreate' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tag' }

  /points:
    get:
      security: [{ clerkBearer: [] }]
      tags: [points]
      summary: List current user's points
      parameters:
        - name: folderId
          in: query
          required: false
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Point' } }
    post:
      security: [{ clerkBearer: [] }]
      tags: [points]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PointCreate' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Point' }
        '400':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /points/{pointId}/photo-upload:
    post:
      security: [{ clerkBearer: [] }]
      tags: [points]
      summary: Issue signed Cloudinary direct-upload parameters for the one allowed image
      parameters:
        - name: pointId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: contentType
          in: query
          schema: { type: string, example: image/jpeg }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PhotoUpload' }
        '403': { description: Not author }
        '404': { description: Not found }

  /points/{pointId}:
    patch:
      security: [{ clerkBearer: [] }]
      tags: [points]
      summary: Update own point (partial)
      parameters:
        - name: pointId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PointUpdate' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Point' }
        '400':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403': { description: Not author }
        '404': { description: Not found }

  /points/{pointId}/comments:
    get:
      security: [{ clerkBearer: [] }]
      tags: [points]
      parameters:
        - name: pointId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Comment' } }
    post:
      security: [{ clerkBearer: [] }]
      tags: [points]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string, minLength: 1, maxLength: 4000 }
      responses:
        '201': { description: Created }
        '401': { description: Unauthorized (guest) }

  /points/{pointId}/rating:
    put:
      security: [{ clerkBearer: [] }]
      tags: [points]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value: { type: integer, minimum: 1, maximum: 5 }
      responses:
        '200': { description: OK }
        '401': { description: Unauthorized }

# Additional routes (favorites, folders, groups, tags) to be expanded
# in the same file as implementation progresses; version bump on breaking change.